+++ app/src/main/java/com/iandp/happy/activity/ShopDetailActivity.java
@@ -29,12 +29,12 @@ public class ShopDetailActivity extends AppCompatActivity {
 
         setToolbar();
 
-        int idShop = getIntent().getIntExtra(DATA_ID_SHOP, -1);
+        long idShop = getIntent().getLongExtra(DATA_ID_SHOP, -1);
         replaceContentFragment(idShop);
 
     }
 
-    protected void replaceContentFragment(int id) {
+    protected void replaceContentFragment(long id) {
         FragmentManager fragmentManager = getSupportFragmentManager();
         Fragment fragmentCurrent = fragmentManager.findFragmentById(R.id.content_frame);
